+++ src/models.rs
@@ -55,7 +55,7 @@ impl<T> MoneroResult<T> {
 }
 
 /// Return type of daemon `get_block_template`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct BlockTemplate {
     pub blockhashing_blob: HashString<Vec<u8>>,
     pub blocktemplate_blob: HashString<Vec<u8>>,
@@ -126,8 +126,32 @@ pub struct BlockHeaderResponse {
     pub timestamp: DateTime<Utc>,
 }
 
-/// Return type of daemon RPC `get_transactions`.
 #[derive(Clone, Debug, Serialize, Deserialize)]
+pub(crate) struct GenerateBlocksResponseR {
+    pub height: u64,
+    pub blocks: Option<Vec<HashString<BlockHash>>>,
+}
+
+/// Return type of regtest daemon RPC `generate_blocks`
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct GenerateBlocksResponse {
+    pub height: u64,
+    pub blocks: Option<Vec<BlockHash>>,
+}
+
+impl From<GenerateBlocksResponseR> for GenerateBlocksResponse {
+    fn from(v: GenerateBlocksResponseR) -> Self {
+        let GenerateBlocksResponseR { height, blocks } = v;
+
+        Self {
+            height,
+            blocks: blocks.map(|vec| vec.into_iter().map(|b| b.0).collect()),
+        }
+    }
+}
+
+/// Return type of daemon RPC `get_transactions`.
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct TransactionsResponse {
     pub credits: u64,
     pub top_hash: String,
@@ -140,7 +164,7 @@ pub struct TransactionsResponse {
 }
 
 /// Sub-type of [`TransactionsResponse`]'s return type of daemon RPC `get_transactions`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Transaction {
     pub as_hex: String,
     pub as_json: Option<String>, // needs to be parsed as JsonTransaction, but is received as a string
@@ -162,7 +186,7 @@ pub struct JsonTransaction {
 }
 
 /// Sub-type of [`BalanceData`]'s return type of wallet `get_balance`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct SubaddressBalanceData {
     pub address: Address,
     pub address_index: u32,
@@ -175,7 +199,7 @@ pub struct SubaddressBalanceData {
 }
 
 /// Return type of wallet `get_balance`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct BalanceData {
     /// Balance amount of account queried.
     #[serde(with = "amount::serde::as_pico")]
@@ -191,7 +215,7 @@ pub struct BalanceData {
 }
 
 /// Argument type of wallet `transfer`.
-#[derive(Copy, Clone, Debug)]
+#[derive(Copy, Clone, Debug, PartialEq)]
 pub enum TransferPriority {
     Default,
     Unimportant,
@@ -208,13 +232,13 @@ pub struct TransferData {
     pub fee: Amount,
     pub tx_blob: HashString<Vec<u8>>,
     pub tx_hash: HashString<CryptoNoteHash>,
-    pub tx_key: HashString<CryptoNoteHash>,
+    pub tx_key: HashString<Vec<u8>>,
     pub tx_metadata: HashString<Vec<u8>>,
     pub unsigned_txset: HashString<Vec<u8>>,
 }
 
 /// Sub-type of [`AddressData`]'s return type of wallet `get_address`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct SubaddressData {
     pub address: Address,
     pub address_index: u32,
@@ -223,7 +247,7 @@ pub struct SubaddressData {
 }
 
 /// Return type of wallet `get_payments`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Payment {
     pub payment_id: HashString<PaymentId>,
     pub tx_hash: HashString<CryptoNoteHash>,
@@ -245,7 +269,7 @@ pub struct WalletCreation {
 }
 
 /// Return type of wallet `get_address`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct AddressData {
     /// Address of the account queried.
     pub address: Address,
@@ -262,13 +286,13 @@ pub enum TransferType {
 }
 
 /// Return type of wallet `incoming_transfers`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct IncomingTransfers {
     pub transfers: Option<Vec<IncomingTransfer>>,
 }
 
 /// Sub-type of [`IncomingTransfers`]. Represent one incoming transfer.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct IncomingTransfer {
     #[serde(with = "amount::serde::as_pico")]
     pub amount: Amount,
@@ -341,12 +365,14 @@ pub struct GenerateFromKeysArgs {
     pub address: Address,
     pub spendkey: Option<monero::PrivateKey>,
     pub viewkey: monero::PrivateKey,
+    // TODO it seems this argument is really optional, although the doc at
+    // `https://www.getmonero.org/resources/developer-guides/wallet-rpc.html#generate_from_keys` does not mention it
     pub password: String,
     pub autosave_current: Option<bool>,
 }
 
 /// Return sub-type of wallet `get_accounts`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct GotAccount {
     pub account_index: u32,
     #[serde(with = "amount::serde::as_pico")]
@@ -366,7 +392,7 @@ pub struct RefreshData {
 }
 
 /// Return type of wallet `get_accounts`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct GetAccountsData {
     pub subaddress_accounts: Vec<GotAccount>,
     #[serde(with = "amount::serde::as_pico")]
@@ -464,7 +490,7 @@ impl<'de> Deserialize<'de> for TransferHeight {
 }
 
 /// Return type of wallet `get_transfer` and `get_transfers`.
-#[derive(Clone, Debug, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Deserialize)]
 pub struct GotTransfer {
     /// Public address of the transfer.
     pub address: Address,
@@ -510,7 +536,7 @@ pub struct SignedTransferOutput {
 
 /// Used to export and import signed key images. Return type of wallet `export_key_images` and
 /// argument type of wallet `import_key_images`.
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct SignedKeyImage {
     /// The key image.
     pub key_image: Vec<u8>,
@@ -519,7 +545,7 @@ pub struct SignedKeyImage {
 }
 
 /// Return type of wallet `import_key_images`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct KeyImageImportResponse {
     pub height: u64,
     /// Amount spent from key images.
@@ -605,5 +631,30 @@ mod tests {
 
         assert_de_tokens(&confirmed, &[Token::U64(10)]);
         assert_de_tokens(&in_pool, &[Token::U64(0)]);
+    }
+
+    #[test]
+    fn generate_blocks_response_from_generate_blocks_response_r() {
+        let gbrr = GenerateBlocksResponseR {
+            height: 10,
+            blocks: None,
+        };
+        let expected_gbr = GenerateBlocksResponse {
+            height: 10,
+            blocks: None,
+        };
+        assert_eq!(GenerateBlocksResponse::from(gbrr), expected_gbr);
+
+        let block_hash = BlockHash::zero();
+
+        let gbrr = GenerateBlocksResponseR {
+            height: 10,
+            blocks: Some(vec![HashString(block_hash)]),
+        };
+        let expected_gbr = GenerateBlocksResponse {
+            height: 10,
+            blocks: Some(vec![block_hash]),
+        };
+        assert_eq!(GenerateBlocksResponse::from(gbrr), expected_gbr);
     }
 }